main페이지/유저 로직변경

+++ board/views.py
@@ -6,6 +6,10 @@ from .models import Post, Member, Comment, Postfile
 
 # Create your views here.
 def postlist(request):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
     post_list = Post.objects.all()
     count = len(post_list)
     
@@ -17,6 +21,10 @@ def postlist(request):
     
 
 def post(request, post_no):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
     if request.method == 'GET':
         post = Post.objects.get(post_no=post_no)
         comments = Comment.objects.filter(post_no=post_no)
@@ -50,6 +58,11 @@ def post(request, post_no):
 
 # 작업중
 def write(request):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
+    
     if request.method == 'POST':
         save_post_title = request.POST.get('postname')
         save_post_detail = request.POST.get('contents')
@@ -88,6 +101,10 @@ def write(request):
     
 
 def update(request, post_no):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
     post = Post.objects.get(post_no=post_no)
     if request.method == "POST":
         post.post_title = request.POST['postname']
@@ -107,6 +124,10 @@ def update(request, post_no):
 
 
 def delete(request, post_no):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
     post = Post.objects.get(post_no=post_no)
     post.delete()
     return postlist(request)
@@ -118,6 +139,10 @@ def delete(request, post_no):
 
 
 def comment_write(request, post_no):
+    s_id = request.session.get('s_id') # session id 유무 체크
+    if s_id==None:
+        return redirect("/main/login/")
+    
     if request.method == 'POST':
         now = datetime.datetime.now()
         save_member_id = request.session.get('s_id')
